FARRAGO: add explicit casts to make the Eclipse compiler happy
[git-p4: depot-paths = "//open/dt/dev/": change = 6298]

+++ farrago/src/com/lucidera/lcs/LcsIndexIntersectRel.java
@@ -126,8 +126,8 @@ class LcsIndexIntersectRel extends FennelMultipleRel
         names[inputs.length + 1] = "rowLimitParamId";
         pw.explain(
             this, names, new Object[] { 
-                (startRidParamId == null) ? 0 : startRidParamId,
-                (rowLimitParamId == null) ? 0 : rowLimitParamId });
+                (startRidParamId == null) ? (Integer)0 : startRidParamId,
+                (rowLimitParamId == null) ? (Integer)0 : rowLimitParamId });
     }
 
     // implement RelNode
